Clear rectangular area
Clear a rectangular area in the canvas starting from the top-left coordinates (0,0), that is, the first two parameters of the clearRect() method and then create a rectangle with its height and width equal to that of the canvas, that is, the last two parameters of the clearRect() method.

Draw an image onto the canvas using the image stored in the image object starting from the coordinates (0,0).