Responsive font size
Set the the font size that is 4% of the average of the image height and width.
The canvas creates a baseline at the location specified for the text. 
By setting the textBaseline to top, the baseline will be on top of the text, that is, that text will be rendered below the baseline